simpify capacity service: accumulate beds for ward the same way as for hosptials and departments
